new master page layouts, IE flexbox fixes, color themeing process
This is a broad commit of changes that was mainly due to changing the
way our Master Page is setup in Kentico.

We did not want to maintain multiple “master pages” in the CMS, rather
we created one page that contained flexible Web Part “zones” that allow
us to add or remove modules based on the content patterns needed.

And IE still has flexbox bugs. Go figure.